Pull animation code out of yeti
